Status method on Function
Moved the standalone fn into the method.
